[Tests] Moved all tests into /tests directory.
