+++ tests/controllers/LoginController.test.ts
@@ -1,16 +1,16 @@
 /*
- * Copyright (c) 2020. This code created and belongs to Atlas render manager project.
+ * Copyright (c) 2021. This code created and belongs to Atlas render manager project.
  * Owner and project architect: Danil Andreev | danssg08@gmail.com |  https://github.com/DanilAndreev
  * Project: atlas-core
- * File last modified: 11/12/20, 5:25 PM
+ * File last modified: 13.11.2020, 20:11
  * All rights reserved.
  */
 
-import Controller from "../core/Controller";
-import LoginController from "./LoginController";
+import Controller from "../../src/core/Controller";
+import LoginController from "../../src/controllers/LoginController";
 import * as request from 'supertest'
-import Server from "../core/Server";
-import {config} from "../config";
+import Server from "../../src/core/Server";
+import {config} from "../../src/config";
 // import {server} from "../index";
 
 describe("controllers -> LoginController", () => {
